+++ lib/dynamic_call.dart
@@ -3,3 +3,4 @@ library dynamic_call;
 export 'package:mercury_client/mercury_client.dart';
 
 export 'src/dynamic_call_base.dart';
+export 'src/dynamic_call_datasource.dart';